Make attachment hook boundaries explicit
Attachment selectors, upload coordination, update and removal workflows, and
mutation-cache plumbing shared one generic module. That made unrelated reasons
to change difficult to locate and let callers bypass any responsibility
boundary by importing individual hooks.

Expose one object from each responsibility module and define its hooks only as
properties of that object. Callers now select the workflow explicitly while
the feature index remains the supported entry point. Keep the API migration
atomic with the split, and isolate the upload error guard so the expanded
feature index does not introduce a runtime import cycle.
