+++ src/App/frontend/src/components/message/ErrorReport.tsx
@@ -5,7 +5,7 @@ import { Flex, FullWidthWrapper } from '@app/form-component';
 import { ErrorSummary } from '@digdir/designsystemet-react';
 
 import classes from 'src/components/message/ErrorReport.module.css';
-import { useAllAttachments } from 'src/features/attachments/hooks';
+import { AttachmentReadModel } from 'src/features/attachments';
 import { FileScanResults } from 'src/features/attachments/types';
 import {
   InstantiationValidation,
@@ -89,7 +89,7 @@ type InfectedFileError = NodeRefValidation & { dataElementId: string };
 
 export function ErrorReportList({ formErrors, taskErrors }: ErrorReportListProps) {
   const getUniqueKeyFromObject = useGetUniqueKeyFromObject();
-  const allAttachments = useAllAttachments();
+  const allAttachments = AttachmentReadModel.useAllAttachments();
 
   const infectedFileErrors: InfectedFileError[] = Object.entries(allAttachments || {}).flatMap(
     ([nodeId, attachments]) => {
