+++ src/App/frontend/src/features/attachments/hooks/index.ts
@@ -1,0 +1,9 @@
+import { AttachmentReadModel } from 'src/features/attachments/hooks/attachmentReadModel';
+import { AttachmentRemoval } from 'src/features/attachments/hooks/attachmentRemoval';
+import { AttachmentUpdate } from 'src/features/attachments/hooks/attachmentUpdate';
+import { AttachmentUpload } from 'src/features/attachments/hooks/attachmentUpload';
+import type { AttachmentActionRemove } from 'src/features/attachments/hooks/attachmentRemoval';
+import type { AttachmentActionUpdate } from 'src/features/attachments/hooks/attachmentUpdate';
+
+export { AttachmentReadModel, AttachmentRemoval, AttachmentUpdate, AttachmentUpload };
+export type { AttachmentActionRemove, AttachmentActionUpdate };
