+++ src/App/frontend/src/layout/FileUpload/Error/FailedAttachments.tsx
@@ -5,8 +5,12 @@ import { XMarkIcon } from '@navikt/aksel-icons';
 import { isAxiosError } from 'axios';
 
 import { getApplicationMetadata } from 'src/features/applicationMetadata';
-import { type IFailedAttachment, isDataPostError } from 'src/features/attachments';
-import { useDeleteFailedAttachment, useFailedAttachmentsFor } from 'src/features/attachments/hooks';
+import {
+  AttachmentReadModel,
+  AttachmentUpload,
+  type IFailedAttachment,
+  isDataPostError,
+} from 'src/features/attachments';
 import { Lang } from 'src/features/language/Lang';
 import { useLanguage } from 'src/features/language/useLanguage';
 import { getValidationIssueMessage } from 'src/features/validation/backendValidation/backendValidationUtils';
@@ -15,8 +19,8 @@ import { isRejectedFileError } from 'src/layout/FileUpload/RejectedFileError';
 import { useIndexedId } from 'src/utils/layout/DataModelLocation';
 
 export function FailedAttachments({ baseComponentId }: { baseComponentId: string }) {
-  const failedAttachments = useFailedAttachmentsFor(baseComponentId);
-  const deleteFailedAttachment = useDeleteFailedAttachment();
+  const failedAttachments = AttachmentReadModel.useFailedAttachmentsFor(baseComponentId);
+  const deleteFailedAttachment = AttachmentUpload.useDeleteFailedAttachment();
   const indexedId = useIndexedId(baseComponentId);
 
   return failedAttachments.length > 0 ? (
