+++ src/App/frontend/src/layout/FileUpload/FileUploadComponent.tsx
@@ -8,7 +8,7 @@ import { CloudUpIcon } from '@navikt/aksel-icons';
 import cn from 'classnames';
 
 import { Label } from 'src/components/label/Label';
-import { useAddRejectedAttachments, useAttachmentsFor, useAttachmentsUploader } from 'src/features/attachments/hooks';
+import { AttachmentReadModel, AttachmentUpload } from 'src/features/attachments';
 import { Lang } from 'src/features/language/Lang';
 import { useLanguage } from 'src/features/language/useLanguage';
 import { useGetOptions } from 'src/features/options/useGetOptions';
@@ -50,9 +50,9 @@ export function FileUploadComponent({
 
   const [showFileUpload, setShowFileUpload] = React.useState(false);
   const mobileView = useIsMobileOrTablet();
-  const attachments = useAttachmentsFor(baseComponentId);
-  const addRejectedAttachments = useAddRejectedAttachments();
-  const uploadAttachments = useAttachmentsUploader();
+  const attachments = AttachmentReadModel.useAttachmentsFor(baseComponentId);
+  const addRejectedAttachments = AttachmentUpload.useAddRejectedAttachments();
+  const uploadAttachments = AttachmentUpload.useAttachmentsUploader();
   const navigation = useNavigation();
   const { langAsString } = useLanguage();
