+++ src/App/frontend/src/layout/FileUpload/FileUploadTable/FileTableButtons.tsx
@@ -5,8 +5,7 @@ import { PencilIcon, TrashIcon } from '@navikt/aksel-icons';
 
 import { DeleteWarningPopover } from 'src/features/alertOnChange/DeleteWarningPopover';
 import { useAlertOnChange } from 'src/features/alertOnChange/useAlertOnChange';
-import { isAttachmentUploaded } from 'src/features/attachments';
-import { useAttachmentsRemover } from 'src/features/attachments/hooks';
+import { AttachmentRemoval, isAttachmentUploaded } from 'src/features/attachments';
 import { Lang } from 'src/features/language/Lang';
 import { useLanguage } from 'src/features/language/useLanguage';
 import classes from 'src/layout/FileUpload/FileUploadTable/FileTableRow.module.css';
@@ -36,7 +35,7 @@ export function FileTableButtons({
   const showEditButton = hasTag && !editWindowIsOpen && !readOnly;
   const { langAsString } = useLanguage();
   const { index, setEditIndex, editIndex } = useFileTableRow();
-  const removeAttachment = useAttachmentsRemover();
+  const removeAttachment = AttachmentRemoval.useAttachmentsRemover();
   const nodeId = useIndexedId(baseComponentId);
 
   // Edit button
