+++ src/App/frontend/src/layout/FileUpload/Summary/summary.ts
@@ -1,6 +1,6 @@
-import { useAttachmentsFor } from 'src/features/attachments/hooks';
+import { AttachmentReadModel } from 'src/features/attachments';
 import type { IAttachment } from 'src/features/attachments';
 
 export function useUploaderSummaryData(baseComponentId: string): IAttachment[] {
-  return useAttachmentsFor(baseComponentId);
+  return AttachmentReadModel.useAttachmentsFor(baseComponentId);
 }
