+++ src/App/frontend/src/layout/FileUploadWithTag/index.tsx
@@ -1,7 +1,7 @@
 import React, { forwardRef, type JSX } from 'react';
 
 import { getApplicationMetadata } from 'src/features/applicationMetadata';
-import { useAttachmentsFor } from 'src/features/attachments/hooks';
+import { AttachmentReadModel } from 'src/features/attachments';
 import { attachmentSelector, makeAttachmentNode } from 'src/features/attachments/tools';
 import { AttachmentSummaryComponent2 } from 'src/layout/FileUpload/AttachmentSummaryComponent2';
 import { FileUploadComponent } from 'src/layout/FileUpload/FileUploadComponent';
@@ -40,7 +40,7 @@ export class FileUploadWithTag extends FileUploadWithTagDef implements ValidateC
   }
 
   useDisplayData(baseComponentId: string): string {
-    const attachments = useAttachmentsFor(baseComponentId);
+    const attachments = AttachmentReadModel.useAttachmentsFor(baseComponentId);
     return attachments.map((a) => a.data.filename).join(', ');
   }
